requirements.txt
alembic
uvicorn
fastapi
pydantic-settings
psycopg2
python-jose
cloudinary
python-multipart
email-validator
passlib
bcrypt==4.0.1
pandas 
numpy
scikit-learn
matplotlib
seaborn
joblib
python-dotenv
jinja2